Add files via upload
Simple game developed in java . where we have player 1 and player 2 . If the player 1 shoot by Gun 1 it's health will reduce 30 and if it shoot by Gun 2 then his health will reduce by 50 . 

For Player 2 we are provide 1 armour option which will somehow protect him from gun. If Player 2 Shoot by Gun 1 and it's armour is on then his health will reduce to 20 otherwise reduce to 30 .  If Player 2 Shoot by Gun 1 and it's armour is on then his health will reduce to 40 otherwise reduce to 50 . 

Player can also take a heal which will make their health again to 100. 

for restriction of health provided by user we allow only the health between 1 to 100.

once the health become 0 then player is consider as Dead.